+++ google/google_cloud_sql/CHANGELOG.md
@@ -1,4 +1,13 @@
-Google Cloud DNS Plugin changelog
+Google Cloud SQL Plugin changelog
+
+v1.2
+-----
+- Fixing filter alias
+- Adding get_replica, delete_replica to instance to allow you to delete the replica
+- Adding patch to instance
+- Adding restore_backup action
+- Adding restore/enable_ha/terminate for ha pairs to test cat
+
 v1.1
 -----
 - Adding in backup_runs resource